# Handle structured body and attributes