--- a/log/internal/slog.go
+++ b/log/internal/slog.go
@@ -7,7 +7,6 @@ import (
 	"context"
 	"fmt"
 
-	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/log"
 
 	"golang.org/x/exp/slog"
@@ -24,7 +23,7 @@ func (h *slogHandler) Handle(ctx context.Context, r slog.Record) error {
 
 	record.SetTimestamp(r.Time)
 
-	record.SetBody(r.Message)
+	record.SetBody(log.StringValue(r.Message))
 
 	lvl := convertLevel(r.Level)
 	record.SetSeverity(lvl)
@@ -57,29 +56,29 @@ func convertLevel(l slog.Level) log.Severity {
 	return log.Severity(l + 9)
 }
 
-func convertAttr(attr slog.Attr) attribute.KeyValue {
+func convertAttr(attr slog.Attr) log.KeyValue {
 	val := convertValue(attr.Value)
-	return attribute.KeyValue{Key: attribute.Key(attr.Key), Value: val}
+	return log.KeyValue{Key: attr.Key, Value: val}
 }
 
-func convertValue(v slog.Value) attribute.Value {
+func convertValue(v slog.Value) log.Value {
 	switch v.Kind() {
 	case slog.KindAny:
-		return attribute.StringValue(fmt.Sprintf("%+v", v.Any()))
+		return log.StringValue(fmt.Sprintf("%+v", v.Any()))
 	case slog.KindBool:
-		return attribute.BoolValue(v.Bool())
+		return log.BoolValue(v.Bool())
 	case slog.KindDuration:
-		return attribute.Int64Value(v.Duration().Nanoseconds())
+		return log.Int64Value(v.Duration().Nanoseconds())
 	case slog.KindFloat64:
-		return attribute.Float64Value(v.Float64())
+		return log.Float64Value(v.Float64())
 	case slog.KindInt64:
-		return attribute.Int64Value(v.Int64())
+		return log.Int64Value(v.Int64())
 	case slog.KindString:
-		return attribute.StringValue(v.String())
+		return log.StringValue(v.String())
 	case slog.KindTime:
-		return attribute.Int64Value(v.Time().UnixNano())
+		return log.Int64Value(v.Time().UnixNano())
 	case slog.KindUint64:
-		return attribute.Int64Value(int64(v.Uint64()))
+		return log.Int64Value(int64(v.Uint64()))
 	default:
 		panic(fmt.Sprintf("unhandled attribute kind: %s", v.Kind()))
 	}