--- a/log/severity.go
+++ b/log/severity.go
@@ -1,10 +1,6 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-// Copyright 2022 The Go Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style
-// license that can be found in the LICENSE file.
-
 package log // import "go.opentelemetry.io/otel/log"
 
 // Severity represents a log record severity (also known as log level).